refactor(frontend): update contextVar dialogs
